requirements.txt
click
pytest
pyyaml
pytest-xdist
jinja2
rich
cfn-flip